requirements.txt
streamlit>=1.36.0
openai>=1.51.0
python-dotenv>=1.0.1
PyYAML>=6.0.2
Pillow>=10.4.0
reportlab>=4.2.2
pandas>=2.2.2